HZN-1520: Move geocoder service to its own module
