+++ features/geocoder/service/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,30 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0"
+           xsi:schemaLocation="
+http://www.osgi.org/xmlns/blueprint/v1.0.0 https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0 http://aries.apache.org/schemas/blueprint-cm/blueprint-cm-1.3.0.xsd
+http://aries.apache.org/blueprint/xmlns/blueprint-ext/v1.5.0 http://aries.apache.org/schemas/blueprint-ext/blueprint-ext-1.5.xsd">
+
+    <cm:property-placeholder id="geocoderConfig" persistent-id="org.opennms.features.geocoder" update-strategy="reload">
+        <cm:default-properties>
+            <cm:property name="activeGeocoderId" value="" /> <!-- by default no geocoder is used -->
+        </cm:default-properties>
+    </cm:property-placeholder>
+
+    <reference id="configAdmin" interface="org.osgi.service.cm.ConfigurationAdmin" />
+
+    <bean id="geocoderServiceManagerConfig" class="org.opennms.features.geocoder.GeocoderServiceManagerConfiguration">
+        <property name="activeGeocoderId" value="${activeGeocoderId}" />
+    </bean>
+
+    <bean id="geocoderServiceManager" class="org.opennms.features.geocoder.service.DefaultGeocoderServiceManager">
+        <argument ref="blueprintBundleContext" />
+        <argument ref="configAdmin" />
+        <argument ref="geocoderServiceManagerConfig"/>
+    </bean>
+    <service ref="geocoderServiceManager" interface="org.opennms.features.geocoder.GeocoderServiceManager">
+        <service-properties>
+            <entry key="registration.export" value="true"/>
+        </service-properties>
+    </service>
+</blueprint>
